Making the redirect to the management console work from within docker compose.
This makes use of a construct called host.docker.internal I think this is cross platform but we will need to verify this.
The host docker internal directive is what allows bridging the network to the host network without leveraging something like NAT.

Longer term we may want to think about HAPROXY or NGinx or one of those.

I also made the containers depend on one another.  So the iridium container now depends oon the maria db container.  Without this on my machine the api was coming up first and it was causing exceptions.